Data snapshot overhaul
Major overhaul:
- Splits global variables into process_globals.py
- Splits generalized processing functions into process_utils.py
- Make error handling more robust and include logging for errors
- Update data snapshot processing to include inquiry and denial indices
- Updates version number to 2.0
